Implement deterministic ancestry updates and default v2 engine

+++ Causal_Web/engine/engine_v2/loader.py
@@ -76,7 +76,7 @@ def load_graph_arrays(graph_json: Dict[str, Any]) -> GraphArrays:
         "bit": np.zeros(n_vert, dtype=np.int8),
         "conf": np.zeros(n_vert, dtype=np.float32),
         "fanin": np.zeros(n_vert, dtype=np.int32),
-        "ancestry": np.zeros((n_vert, 4), dtype=np.int32),
+        "ancestry": np.zeros((n_vert, 4), dtype=np.uint64),
         "m": np.zeros((n_vert, 3), dtype=np.float32),
         "rho_mean": np.asarray(
             [nodes[nid].get("rho_mean", 0.0) for nid in nodes], dtype=np.float32
